Create socksserver-with-routing-rules.py
A socks server that I was going to write that allows you to override where requests go to,
supports username/password authentication and has rules on a username/password basis.

Additionally I designed/added a multiplexing protocol that starts after the socks protocol.
This was all fine and dandy until I learned the fact that authentication was supported in browsers
for socks5. This lead to the conclusion that I would need to write an HTTP proxy server and I just
don't feel like doing that. So this remains here as a snippet into what could have been and some good bits of code.