Randomization
Just some code I wrote in csharp and tested with mono. I learned about
generic methods and type arguments and used them to choose an amount
of items at random having the chance and item in a tuple. The commented out
code is from when I was trying to figure it out initially.